Move routes from app.py to webgame module and bluprint
Restructure web organization to utilize Blueprints. Still need to deter-
mine best way to keep scorepad object data when redirecting to different
templates throughout a turn and during the whole game. May need to
utilize sqlite since it can be persistent during session.
