+++ pkg/web/temptest/templates/temptest/test01.html
@@ -1,0 +1,6 @@
+
+
+<h1>This is the teplate in the temptest directory</h1>
+<div>
+<h2>{{ name }}</h2>
+</div>
