+++ pkg/web/webgame/templates/webgame/roll_one.html
@@ -1,0 +1,175 @@
+{% extends "base.html" %}
+
+{% block title %}dice-cvn{% endblock %}
+
+{% block content %}
+
+<div class="jumbotron jumbotron-fluid">
+  <div class="container">
+    <h2 class="display-4">First Roll - webgame blueprint</h2>
+    <!-- <p class="lead">This is a modified jumbotron that occupies the entire horizontal space of its parent.</p> -->
+  </div>
+
+  <div class="container">
+
+    <div class="row">
+      <div class="col-sm">
+      <p class="font-weight-bolder"> --- ONE ---</p>
+      </div>
+      <div class="col-sm">
+      <p class="font-weight-bolder">--- TWO ---</p>
+      </div>
+      <div class="col-sm">
+      <p class="font-weight-bolder">--- THREE ---</p>
+      </div>
+      <div class="col-sm">
+      <p class="font-weight-bolder">--- FOUR ---</p>
+      </div>
+      <div class="col-sm">
+      <p class="font-weight-bolder">--- FIVE ---</p>
+      </div>
+    </div>
+
+    <div class="row">
+
+      <div class="col-sm">
+      <img class="img1" src="{{url_for('static', filename=scorepad.web_dice_list[0])}}" width="100" height="100"/>
+      </div>
+
+      <div class="col-sm">
+      <img class="img1" src="{{url_for('static', filename=scorepad.web_dice_list[1])}}" width="100" height="100"/>
+      </div>
+
+      <div class="col-sm">
+      <img class="img1" src="{{url_for('static', filename=scorepad.web_dice_list[2])}}" width="100" height="100"/>
+      </div>
+
+      <div class="col-sm">
+      <img class="img1" src="{{url_for('static', filename=scorepad.web_dice_list[3])}}" width="100" height="100"/>
+      </div>
+
+      <div class="col-sm">
+      <img class="img1" src="{{url_for('static', filename=scorepad.web_dice_list[4])}}" width="100" height="100"/>
+      </div>
+  </div>
+</div>
+
+<p></p>
+<p></p>
+
+<div class="container">
+
+<h4>Dice to Keep:  ***** TEST *****</h4>
+
+<form method="POST">
+
+  <div class="form-check">
+    {{ dice_hold_web_form.hidden_tag() }}
+    {{ dice_hold_web_form.die1.label }} {{ dice_hold_web_form.die1(class="form-check-input") }}
+
+    {{ dice_hold_web_form.roll_dice(class="btn btn-primary")}}
+    <!-- <input class="form-check-input" type="checkbox" value="" id="die1">
+    <label class="form-check-label" for="defaultCheck1">
+      --- ONE TEST ---
+    </label> -->
+
+  </div>
+</form>
+</div>
+
+
+<p></p>
+<p></p>
+<p></p>
+<p></p>
+<p></p>
+<p></p>
+<p></p>
+<p></p>
+
+<div class="container">
+
+<h4>Dice to Keep:</h4>
+
+</div>
+
+<div class="container">
+
+<div class="row">
+<div class="form-check">
+  <input class="form-check-input" type="checkbox" value="1" id="die-1">
+  <label class="form-check-label" for="defaultCheck1">
+    --- ONE ---
+  </label>
+
+  </div>
+
+  <div class="form-check">
+    <input class="form-check-input" type="checkbox" value="2" id="die-2">
+    <label class="form-check-label" for="defaultCheck1">
+      --- TWO ---
+    </label>
+  </div>
+
+  <div class="form-check">
+    <input class="form-check-input" type="checkbox" value="3" id="die-3">
+    <label class="form-check-label" for="defaultCheck1">
+      --- THREE ---
+    </label>
+  </div>
+
+  <div class="form-check">
+    <input class="form-check-input" type="checkbox" value="4" id="die-4">
+    <label class="form-check-label" for="defaultCheck1">
+      --- FOUR ---
+    </label>
+  </div>
+
+  <div class="form-check">
+    <input class="form-check-input" type="checkbox" value="5" id="die-5">
+    <label class="form-check-label" for="defaultCheck1">
+      --- FIVE ---
+    </label>
+  </div>
+
+</div>
+
+</div>
+
+<p></p>
+
+<div class="container">
+
+      <button class="btn btn-primary" type="submit">Roll Dice</button>
+
+      <a class="btn btn-primary" href="/exit_game">Keep All Dice</a>
+
+</div>
+
+<p></p>
+
+<div class="container">
+<a class="btn btn-danger" href="/exit_game">Exit Game</a>
+</div>
+
+<div class="container">
+
+  <p>
+    scorepad.name : {{scorepad.name}}
+  </p>
+
+  <p>
+    scorepad.web_turn_tracking : {{scorepad.web_turn_tracking}}
+  </p>
+
+  <p>
+    scorepad.track_fours : {{scorepad.track_fours}}
+  </p>
+
+  <p>
+    track_fours : {{ track_fours }}
+  </p>
+
+</div>
+
+{% endblock %}
