+++ pkg/web/webgame/templates/webgame/roll_two.html
@@ -4,7 +4,7 @@
 
 {% block content %}
 
-<h1>Roll # 2 in web/templates</h1>
+<h1>Roll # 2 (In webgame/template)</h1>
 
 <h1>Get dice roll and update/render here</h1>
 <h1>die1: {{die1}}</h1>
